+++ genometools/expression/__init__.py
@@ -1,0 +1,6 @@
+from .gene import ExpGene
+from .genome import ExpGenome
+from .matrix import ExpMatrix
+from .normalize import quantile_normalize
+
+__all__ = ['ExpGene','ExpGenome','ExpMatrix','quantile_normalize']