+++ genometools/misc/__init__.py
@@ -1,0 +1,2 @@
+from genometools.misc.functions import *
+from genometools.misc.log import *